Changing package structure to reflect group name
Changes the package structure to separate by group/version rather than
mutator vs validator. Also moves some code to a new "common" package for
easy re-use.
