+++ pkg/resources/management.cattle.io/v3/podsecurityadmissionconfigurationtemplate/validator.go
